% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_purrs.R
\name{purrs.colors}
\alias{purrs.colors}
\title{Get the cat colored palette}
\usage{
purrs.colors(n = "lily")
}
\arguments{
\item{n}{name of the color palette. Defaults to lily. Available options are: "lily", "binky", "teddy", "ginger", "grumpyCat", "caliente", "disney"}
}
\value{
The color palette of choice (catlette)
}
\description{
This function loads the respective palette and applies the color to the ggplot.
}
