% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_purrs.R
\name{scale_fill_purrs}
\alias{scale_fill_purrs}
\title{Fill your ggplot graphs with colors based on cats}
\usage{
scale_fill_purrs(palette_name = "lily", ...)
}
\arguments{
\item{palette_name}{name of the colour palette}
}
\value{
ggplot colored graph
}
\description{
This function loads the respective palette and applies the fill color to the ggplot.
}
