% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_cat}
\alias{theme_cat}
\title{This function redefines the ggplot theme.
It is based on theme_test(), with Raleway Medium as a title font and IBM Plex Sans as the main font.}
\usage{
theme_cat(
  base_size = 12,
  dark_text = "#1A242F",
  title_font = "Raleway Medium",
  main_font = "IBM Plex Sans"
)
}
\arguments{
\item{base_size}{Base size of the font}

\item{dark_text}{The darkest color of the text. The function will define two lighter colors based on this.}

\item{title_font}{The name of the title font}

\item{main_font}{The name of the main font}
}
\description{
This function redefines the ggplot theme.
It is based on theme_test(), with Raleway Medium as a title font and IBM Plex Sans as the main font.
}
